+++ api/internal/features/container/controller/init.go
@@ -2,10 +2,14 @@ package controller
 
 import (
 	"context"
+	"fmt"
+	"strings"
+
 	"github.com/raghavyuva/nixopus-api/internal/features/deploy/docker"
 	"github.com/raghavyuva/nixopus-api/internal/features/logger"
 	"github.com/raghavyuva/nixopus-api/internal/features/notification"
 	shared_storage "github.com/raghavyuva/nixopus-api/internal/storage"
+	shared_types "github.com/raghavyuva/nixopus-api/internal/types"
 )
 
 type ContainerController struct {
@@ -29,4 +33,21 @@ func NewContainerController(
 		logger:        l,
 		notification:  notificationManager,
 	}
+}
+
+func (c *ContainerController) isProtectedContainer(containerID string, action string) (*shared_types.Response, bool) {
+    details, err := c.dockerService.GetContainerById(containerID)
+    if err != nil {
+        return nil, false
+    }
+    name := strings.ToLower(details.Name)
+    if strings.Contains(name, "nixopus") {
+        c.logger.Log(logger.Info, fmt.Sprintf("Skipping %s for protected container", action), details.Name)
+        return &shared_types.Response{
+            Status:  "success",
+            Message: "Operation skipped for protected container",
+            Data:    map[string]string{"status": "skipped"},
+        }, true
+    }
+    return nil, false
 }
